requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Faker==25.3.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
ibm-cos-sdk==2.13.4
ibm-cos-sdk-core==2.13.4
ibm-cos-sdk-s3transfer==2.13.4
ibm_watsonx_ai==1.0.4
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.1
langchain-core==0.2.1
langchain-text-splitters==0.2.0
langsmith==0.1.63
lomond==0.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
orjson==3.10.3
packaging==23.2
pandas==2.1.4
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.1.0
pydantic==2.7.2
pydantic_core==2.18.3
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
streamlit==1.35.0
tabulate==0.9.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.1.0
yarl==1.9.4
zipp==3.19.0
